requirements.txt
openai>=1.50.0
PyMuPDF>=1.24.0
python-frontmatter>=1.1.0
PyYAML>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=8.0.0
